Rename Layout to LayoutBase and move to a folder
