+++ Builder/sources/Domain/Configuration/Template.cs
@@ -9,6 +9,7 @@ namespace Builder.Domain.Configuration
       public IEnumerable<string> Languages { get; set; }
       public IEnumerable<TemplateHtmlFile> Index { get; set; }
       public IEnumerable<TemplateHtmlFile> Blog { get; set; }
+      public IEnumerable<TemplateHtmlFile> Page { get; set; }
       public IEnumerable<TemplateHtmlFile> Post { get; set; }
       public TemplateLoopHtmlFile Loops { get; set; }
    }
