--- Builder/sources/Domain/LayoutLoop.cs
@@ -1,17 +1,0 @@
-﻿using Builder.Domain.Configuration;
-
-namespace Kennis.Builder.Domain
-{
-    public class LayoutLoop
-   {
-      public string BlogArchive { get; set; }
-      public string BlogCategories { get; set; }
-      public string BlogPostLast10 { get; set; }
-      public string BlogPostLast5 { get; set; }
-      public string BlogPostLast3 { get; set; }
-      public string BlogPosts { get; set; }
-      public string BlogTags { get; set; }
-      public string Menu { get; set; }
-      public string SocialMedia { get; set; }
-   }
-}
