--- Builder/sources/Domain/LayoutTemplate.cs
@@ -1,8 +1,0 @@
-﻿namespace Builder.Domain
-{
-   public class LayoutTemplate
-   {
-      public string Template { get; set; }
-      public List<LayoutTemplatePreprocessed> TemplatesPreprocessed { get; set; }
-   }
-}
