--- Builder/sources/Domain/LayoutTemplatePreprocessed.cs
@@ -1,8 +1,0 @@
-﻿namespace Builder.Domain
-{
-   public class LayoutTemplatePreprocessed
-   {
-      public Guid Id { get; set; }
-      public string Template { get; set; }
-   }
-}
