+++ Builder/sources/Domain/Layouts/LayoutTemplate.cs
@@ -1,0 +1,8 @@
+﻿namespace Builder.Domain.Layouts
+{
+    public class LayoutTemplate
+    {
+        public string Template { get; set; }
+        public List<LayoutTemplatePreprocessed> TemplatesPreprocessed { get; set; }
+    }
+}
