+++ Builder/sources/Domain/Layouts/LayoutTemplatePreprocessed.cs
@@ -1,0 +1,8 @@
+﻿namespace Builder.Domain.Layouts
+{
+    public class LayoutTemplatePreprocessed
+    {
+        public Guid Id { get; set; }
+        public string Template { get; set; }
+    }
+}
